Bind data to grid cells

This function needs a dataset with x and y coordinates. It expects a tibble with columns named 'x' and 'y'. The function will add a column 'cell' to the dataset with the grid cell number. The grid cell number is determined by the cell number in the grid file for the given resolution.

Usage

grid_cell(d, resolution)

Arguments

d

The dataset with x and y coordinates

resolution

The resolution of the grid

Value

The dataset with an additional column with the grid cell number

Details

This function binds the x and y coordinates of a dataset to the grid cell in a map with a specific resolution. The grid cell is determined by the cell number in the grid file.

Examples

data(dclutum)
d <- dclutum
grid_cell(d, 250)
#> # A tibble: 99,980 × 4
#> # Rowwise: 
#>          x       y lutum    cell
#>      <dbl>   <dbl> <dbl>   <dbl>
#>  1 146925. 439590. 44.9   830508
#>  2 127389. 477110. 36     662430
#>  3 208395. 485642.  4.36  624674
#>  4 252786. 476240.  4.50  667412
#>  5 136242. 380278.  2    1095905
#>  6 187202. 594169. 28.8   138509
#>  7 261186. 543939.  4     363925
#>  8 141540  570548.  3.49  243607
#>  9 126528. 471858. 55.9   685947
#> 10 168183. 600756.  2     108193
#> # ℹ 99,970 more rows